Backend/requirements.txt
Django==1.11.4
djangorestframework==3.6.3
isodate==0.5.4
pyparsing==2.2.0
pytz==2017.2
rdflib==4.2.2
SPARQLWrapper==1.8.0
